+++ app/src/main/java/org/simple/clinic/navigation/v2/ActivityResult.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.router.screen
+package org.simple.clinic.navigation.v2
 
 import android.app.Activity
 import android.content.Intent