+++ app/src/main/java/org/simple/clinic/selectcountry/SelectCountryScreenKey.kt
@@ -3,7 +3,7 @@ package org.simple.clinic.selectcountry
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.compat.FullScreenKey
 
 @Parcelize
 class SelectCountryScreenKey : FullScreenKey {