+++ app/src/main/java/org/simple/clinic/util/AndroidExtensions.kt
@@ -7,10 +7,14 @@ import android.content.Intent
 import android.content.res.Configuration
 import android.os.Build
 import android.view.KeyEvent
+import androidx.annotation.AttrRes
+import androidx.annotation.ColorInt
+import androidx.annotation.ColorRes
+import androidx.core.content.ContextCompat
 import io.reactivex.Observable
 import org.simple.clinic.feature.Feature
 import org.simple.clinic.feature.Features
-import org.simple.clinic.router.screen.ActivityResult
+import org.simple.clinic.navigation.v2.ActivityResult
 import java.util.Locale
 
 inline fun Context.wrap(wrapper: (Context) -> Context): Context = wrapper(this)
@@ -73,5 +77,40 @@ inline fun Dialog.onBackPressed(crossinline backPressed: () -> Unit) {
     } else {
       false
     }
+  }
+}
+
+// Source: https://github.com/afollestad/material-dialogs/blob/66582d9993bcf55bfea873b4bf484a429ce3df36/core/src/main/java/com/afollestad/materialdialogs/utils/MDUtil.kt#L103
+@ColorInt
+fun Context.resolveColor(
+    @ColorRes colorRes: Int? = null,
+    @AttrRes attrRes: Int? = null,
+    fallback: (() -> Int)? = null
+): Int {
+  if (attrRes != null) {
+    val a = theme.obtainStyledAttributes(intArrayOf(attrRes))
+    try {
+      val result = a.getColor(0, 0)
+      if (result == 0 && fallback != null) {
+        return fallback()
+      }
+      return result
+    } finally {
+      a.recycle()
+    }
+  }
+  return ContextCompat.getColor(this, colorRes ?: 0)
+}
+
+fun Context.resolveFloat(attrRes: Int, fallback: (() -> Float)? = null): Float {
+  val a = theme.obtainStyledAttributes(intArrayOf(attrRes))
+  try {
+    val result = a.getFloat(0, 0f)
+    if (result == 0f && fallback != null) {
+      return fallback()
+    }
+    return result
+  } finally {
+    a.recycle()
   }
 }